+++ demo-source/src/InstallDemo.css
@@ -1,0 +1,12 @@
+.install-demo {
+	width: 100%;
+	height: 100dvh;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.install-demo .loader-icon {
+	margin: 1rem;
+	height: 3rem;
+}
