--- demo-source/src/LoadDemo.css
@@ -1,12 +1,0 @@
-.load-demo {
-	width: 100%;
-	height: 100dvh;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
-.load-demo .loader-icon {
-	margin: 1rem;
-	height: 3rem;
-}
