+++ demo-source/src/index.js
@@ -7,9 +7,9 @@ import { BrowserRouter } from 'react-router-dom';
 import SelectFramework from './SelectFramework';
 import Embedded from './Embedded';
 import Home from './Home';
-import LoadDemo from './LoadDemo';
+import InstallDemo from './InstallDemo';
 import Editor from './Editor';
-import { onMessage } from './msg-bus';
+import { onMessage } from 'php-cgi-wasm/msg-bus';
 
 navigator.serviceWorker.register(process.env.PUBLIC_URL + `/cgi-worker.js`);
 
@@ -23,34 +23,34 @@ root.render(
 	<BrowserRouter basename={process.env.PUBLIC_URL}>
 		<Routes>
 			<Route path = "/" element = { <Home /> } />
-			<Route path = "/home" element = { <Home /> } />
-			<Route path = "/embedded" element = { <Embedded /> } />
-			<Route path = "/select-framework" element = { <SelectFramework /> } />
-			<Route path = "/load-demo" element = { <LoadDemo /> } />
-			<Route path = "/code-editor" element = { <Editor /> } />
+			<Route path = "/home.html" element = { <Home /> } />
+			<Route path = "/embedded-php.html" element = { <Embedded /> } />
+			<Route path = "/select-framework.html" element = { <SelectFramework /> } />
+			<Route path = "/install-demo.html" element = { <InstallDemo /> } />
+			<Route path = "/code-editor.html" element = { <Editor /> } />
 			<Route
-				path = "/php-wasm/code-editor"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/code-editor' + window.location.search} />}
+				path = "/php-wasm/code-editor.html"
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/code-editor.html' + window.location.search} />}
 			/>
 			<Route
 				path = "/php-wasm/cgi-bin/drupal"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/load-demo?framework=drupal-7'} />}
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/install-demo.html?framework=drupal-7'} />}
 			/>
 			<Route
 				path = "/php-wasm/cgi-bin/cakephp-5"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/load-demo?framework=cakephp-5'} />}
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/install-demo.html?framework=cakephp-5'} />}
 			/>
 			<Route
 				path = "/php-wasm/cgi-bin/codeigniter-4"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/load-demo?framework=codeigniter-4'} />}
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/install-demo.html?framework=codeigniter-4'} />}
 			/>
 			<Route
 				path = "/php-wasm/cgi-bin/laminas-3"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/load-demo?framework=laminas-3'} />}
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/install-demo.html?framework=laminas-3'} />}
 			/>
 			<Route
 				path = "/php-wasm/cgi-bin/laravel-11"
-				element = { <Navigate to = {process.env.PUBLIC_URL + '/load-demo?framework=laravel-11'} />}
+				element = { <Navigate to = {process.env.PUBLIC_URL + '/install-demo.html?framework=laravel-11'} />}
 			/>
 		</Routes>
 	</BrowserRouter>
