+++ docs/code-editor.html
@@ -1,0 +1,52 @@
+<!doctype html><html lang="en"><head><meta charset="utf-8"/><link rel="icon" href="/php-wasm/favicon.ico"/><meta name="viewport" content="width=650,user-scalable=no"/><meta name="theme-color" content="#000000"/><meta name="description" content="Web site created using create-react-app"/><link rel="apple-touch-icon" href="/php-wasm/logo192.png"/><link rel="manifest" href="/php-wasm/manifest.json"/><title>php-wasm</title><script type="text/javascript" src="https://unpkg.com/curvature@0.0.68-h/dist/curvature.js"></script><script defer="defer" src="/php-wasm/static/js/main.1be69e15.js"></script><link href="/php-wasm/static/css/main.afc6f3f6.css" rel="stylesheet"></head><body><noscript>You need to enable JavaScript to run this app.</noscript><div id="root"></div></body></html><!--
+sssooo+++++++///////++++//:----:::::///////:::::::///+ossssssyyyhdddddmmmmNNmo`````````````````````-
+soooo+//////::::::::////:-....-----::::::----------:/+ossyyyyhdhddmmmmNNMNNNN:                     .
+soo++//:::---------:://::-----..------.-.........--:/+ossyyhddmmddmmNNNNMMNNm.                     .
+ssoo+//:-----.....-:::::..................```.`..-::/+osyhdhdddmmmmmNMNMMNNNm`                     .
+ssoo++:::----.....--:/:--.....--...--...`........--:/+osyhdddddmmmddNNNMNNNNd`                     .
+sysso+///::--.....-::/:--....-------.-..........---:/+osyhhddmmNmmmmNNNNNNNNs                      .
+syyssoo++++/+/::--:://:-...---::::::----------:::::///+syhmmNNNNmNNNNNMNNNNNs                      .
+syssssoooooooo+/////++/:-.-:://///////::::--:::::///+osydddddmNNNNNNNNMNNNNN-                      .
+yyssssssssooo+++++++o+/:--:/+//++///////::----:///++osyyysoshdNNNNNNNNNNNNNs`                      .
+yhyyssssossoo+++ooo+++/:-::////+++++////:::-::://+oooo++/:/+ohmNNNNNNNNNNNh`                       .
+yhyyssoooooo++/++////+:--::://+++++/::------::://++/////+///+sdNNNNNNNNmNd.                        .
+ssysssooooo+////::::::-----:///:::---.....----::::::/++////++shdmmNNNNNNm.`                        .
+sooo+//+++////::::-.......--:--....````......--:/+ooo+:--::/oshmNNNNNNNNo`                         .
+sssso++/:/:::::::--.`````.....````````..`.--:////:::--..--:+syhmNNNNNNms`                          .
+///+++++////:/::::-``````````````..---:-::::://:::-:///++++oshmNNNNmNho.                           .
++yhh/:+/-:/++/+///:-........```.-://+++/::+yhoosh+-ohhdmmmmmmmmmmNNNy-`-                           .
+hds/+/++SeanMorris/:-.-.-...``.-:+syhdmy/::oh++oyosyydmNNmmmmmmmmmNN/` `.                          .
+hhhhyhhmmNNNNNNNNmhs+///:-..``-/ydNNNNNNmdyso++/+osyhhhhhhhhdmmmmmNd`                              .
+yyysyyhhdmNNNNMMNNNmhsso+/::-.+h-PHP-wasm#so+/+//+oosssooosssymmmNN+`                              .
+ooooossyhdddmNNNNNNmdyyso+++//yyhdmddhhyyso+//+oooooooosssyyhdhmNNm:.`                             .
+ssssssyhhhhdmNmmNNmdhyyssssooshyyyhdddhyso+++++++++oo++oossyhmddNNy`                               .
+ssyyyysssshdddmNNmmdhyyssssoosyyso+ooosyssoo+++/////+++ossyydmmmNN:                                .
+sssossyyhhyysmNmdddhhyyssssoooshyo//++/+++++++++++ooooosssyhdmNNNd`                                .
+ssssssyyyssyymmdhhhyyyssssoo+oosyyyyso++++++++oosssssssssyhddmNNNo                                 .
+syysssyyyyyyhddhyyyyssssoo++///++syyssoo+/+++++ossssssssyyhdmNNNN:                                 .
+sssoooooossyhdddhhhhyssooo+/////++++//:::--:--:///+oosssyhdmNNNNm.                                 .
++++/++++osssyyhmddddhyssso+++++++/::::------......--:/+syhdmNNNNd:.```      ````                   .
+++ooooo+ohmNmmddmmddhhyyssooooso++oso+-....----..```.-:+yhdmNNNMNNmdhyyyyssshddhso/--..`           .
+oooooo+ooodmmmmddmmdhhysssssssssoyso+:---....-:/:-.``.-/shdmNNNMMMMMMNNNNNNNNMMNMNNNNmdhys+/:-.``  .
+syysssssosyhhdddddddhyysssossssssso//::://////++++/:---:shdmNNNMMMNNNNNMMMNNMNMMMNMMMMMNNNNNNmdhysoo
+ssssssysssssshdddddhyyssoooooooooo/:///+ooooooooo+///o/+ydmmNNNMMMNMNNMMMMMMMMMMMMMMMMMNNMNNNNNNNNNd
+hys++oyyyssoooyhhhhyyso++/+oooo++///+++ooooo++ss////+soshmNNNMMMMMMMNNMMMMMMMMMMMMMMMMMNMMNNNNNNNNNm
+ydmdyooosssooooosyysso+/::::+++//++++oo++//+sys++++oosyhmNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNNMNNNm
+yhdmmmdhyoo+++oo++++/:.....-://++o++++/+oyhdyoosssssyhhmNNMMMMMMMNNMMMMMMMMNMMMMMMMMMMMMMMMNMMMMMMMm
+hhyhmNNNNdyysso++///:-.`````--:///++osydmmdyossssssyddmNMMMMMMMMMMNMMMMMMNNMMMMNMMMMMMMMMMMMMMMMMNMm
+hhysydNNNN+.-ossysso+//:/::://+ooo/:mNmmmhsssyssyyhdmmNNMMMMMMMMMMNMMMMMMNNNMMNNMMMMMMNMMNNNNNNNMNMm
+hhyyyydmmmh-./``.///++oo+//::-.`-::smmmhsssyyyyyhhdNNNMMNNMMMMMMNMMMMMMNNNNMMMNNMMMMNNNMNNNNNNNNMMMm
+hdhyyyyhmmhyyy:.`-`   `.   `-.`.shhdddyosyyyhyyhdmmNNNMNNNMMMMMNNMMNMMMMNNMNMMMMMMNNNMMMNNNmNNNNNNNm
+dmdhyyssyhdhyyyysyo+//o+/:/oyysshdmdys+syyyyyhhdmmNmNMNNNNMMMMMNMNMMMMMNNNMMMMMMMMMMMMMNNNNNNNNNNMMm
+mmmdhyysssyhhyyyssssssyyyyyhyyyyhhyo+oosysyshddmmmmmNMNNNNMMMNNNMMMMMMNNNNMMMMMNMNMMNMNNmmmNNNNNNNNm
+mNNmdhhyyyyyssyyssssossysssysyyso+++osysssshhmNmmmmNNNNNNMMMNNMMMMMMMMNMMMMMMMNNNNMMNNNNmmmNNNNNNNMm
+mNNmmdhdhhyyysossssysyyyyysooo+++osssssyyyhdNNddmdmNMMNNMMMNNNMMMNMMMMNMMNNMNNNNNNNNMNNmmmNNNNNNNNMm
+dmNmmddhhhhhyyssssosooooooooooosoossyyyyyhdmmdddmmmNMNNNMMNNNNNMMMMMMMNMMMNNNNNNNNNNNNmmmmNNNNNNNMNm
+hddmmmddhdhhhyyssssssssysssoooosyyyyyyhhhdddhmmdmmNMMNMNNNNmNNNMMMMNNNNMNMNNNNNNMNMNNmmmmmmNNNNNMMMm
+ddhhddmdddhyyyyyyyssysssssssssyyyyyyhhddddhdmmNNNNNNNMMMNNmmNNNMNMMNNNNNNNNNNNNNNNNNmmdmmmmNNNNNNMMm
+hdhhyydmmdhhyhyyyyyssyysssssyysyyyhhhdmdhydNNNNNNMMNNMNNNmmmNNNMNNNNmmNNNNNNNNNNNNNNmmddmmNNmNNNMMNm
+yhyysosshhddhdhyyyssssyyyyyyyysyyyhdmdyyhdNNNNNMMNNNNMMNmmmmmNNMNNNmmmNNNNNNNNNNNNNmmddmmmNNmNNMMNNd
+yssooooo+oydddyhyhhyyyyysyhyhhyhhyyhssshdNNNMMMMNNNNMNNNmdhmmNNMNmmddmmNNNNNNNNMNNNmdhddmmNmNNNMNNNh
+sso+++++++++oyhdddddyhhhyyhhdddhho++oyhdmNNNMMMNNNNNNNNmdhhmNNNNNmhddmmNNNmNNNNNNNmddddmmmmNNNNNNNmh
+sso++++++///++//oyyyyhhhyyyyssoo+++syhhhdmNNNNmmmmmmmmdhhyhmNNNmddhhddmmmmmmmmNmNmmdhhhdddmmmmNNmmdy
+-->
